refactor: 用菜单插槽承载卡片业务操作
